refactor: unify process descriptor handling and improve logging
- Replaced legacy file descriptor installation with a unified descriptor table for standard input/output.
- Updated process lifecycle management to enhance logging for child process states and reaping.
- Removed unused filesystem includes and adjusted comments for clarity.
- Improved error handling in process loading from ObjectFS.

+++ userland/apps/keynotify/keypress.h
@@ -1,6 +1,9 @@
 #pragma once
 
-typedef unsigned keypress_type_t;
+// Ensure userland struct layout matches kernel (see kernel/src/lib/keyboard.h)
+#include <stdint.h>
+
+typedef uint8_t keypress_type_t;
 typedef keypress_type_t keypress_type;
 
 #define KEYBOARD_KEY_PRESSED ((keypress_type)1)
